Removed enableSchedulerTracing flag and turned tracing on for all builds
