+++ packages/react-reconciler/src/__tests__/ReactHooksWithNoopRenderer-test.internal.js
@@ -60,7 +60,7 @@ describe('ReactHooksWithNoopRenderer', () => {
     ReactFeatureFlags = require('shared/ReactFeatureFlags');
     ReactFeatureFlags.debugRenderPhaseSideEffectsForStrictMode = false;
     ReactFeatureFlags.enableHooks = true;
-    ReactFeatureFlags.enableSchedulerTracing = true;
+
     React = require('react');
     ReactNoop = require('react-noop-renderer');
     SchedulerTracing = require('scheduler/tracing');
