+++ packages/react/src/__tests__/ReactProfiler-test.internal.js
@@ -26,7 +26,6 @@ let resourcePromise;
 
 function loadModules({
   enableProfilerTimer = true,
-  enableSchedulerTracing = true,
   replayFailedUnitOfWorkWithInvokeGuardedCallback = false,
   useNoopRenderer = false,
 } = {}) {
@@ -40,7 +39,6 @@ function loadModules({
   ReactFeatureFlags.debugRenderPhaseSideEffects = false;
   ReactFeatureFlags.debugRenderPhaseSideEffectsForStrictMode = false;
   ReactFeatureFlags.enableProfilerTimer = enableProfilerTimer;
-  ReactFeatureFlags.enableSchedulerTracing = enableSchedulerTracing;
   ReactFeatureFlags.replayFailedUnitOfWorkWithInvokeGuardedCallback = replayFailedUnitOfWorkWithInvokeGuardedCallback;
 
   React = require('react');
@@ -127,1075 +125,1067 @@ const mockDevToolsForTest = () => {
 
 describe('Profiler', () => {
   describe('works in profiling and non-profiling bundles', () => {
-    [true, false].forEach(enableSchedulerTracing => {
-      [true, false].forEach(enableProfilerTimer => {
-        describe(`enableSchedulerTracing:${
-          enableSchedulerTracing ? 'enabled' : 'disabled'
-        } enableProfilerTimer:${
-          enableProfilerTimer ? 'enabled' : 'disabled'
-        }`, () => {
-          beforeEach(() => {
-            jest.resetModules();
+    [true, false].forEach(enableProfilerTimer => {
+      describe(`enableProfilerTimer:${
+        enableProfilerTimer ? 'enabled' : 'disabled'
+      }`, () => {
+        beforeEach(() => {
+          jest.resetModules();
 
-            loadModules({enableSchedulerTracing, enableProfilerTimer});
-          });
+          loadModules({enableProfilerTimer});
+        });
 
-          // This will throw in production too,
-          // But the test is only interested in verifying the DEV error message.
-          if (__DEV__ && enableProfilerTimer) {
-            it('should warn if required params are missing', () => {
+        // This will throw in production too,
+        // But the test is only interested in verifying the DEV error message.
+        if (__DEV__ && enableProfilerTimer) {
+          it('should warn if required params are missing', () => {
+            expect(() => {
               expect(() => {
-                expect(() => {
-                  ReactTestRenderer.create(<React.unstable_Profiler />);
-                }).toThrow('onRender is not a function');
-              }).toWarnDev(
-                'Profiler must specify an "id" string and "onRender" function as props',
-                {withoutStack: true},
-              );
-            });
-          }
-
-          it('should support an empty Profiler (with no children)', () => {
-            // As root
-            expect(
-              ReactTestRenderer.create(
-                <React.unstable_Profiler id="label" onRender={jest.fn()} />,
-              ).toJSON(),
-            ).toMatchSnapshot();
-
-            // As non-root
-            expect(
-              ReactTestRenderer.create(
-                <div>
-                  <React.unstable_Profiler id="label" onRender={jest.fn()} />
-                </div>,
-              ).toJSON(),
-            ).toMatchSnapshot();
+                ReactTestRenderer.create(<React.unstable_Profiler />);
+              }).toThrow('onRender is not a function');
+            }).toWarnDev(
+              'Profiler must specify an "id" string and "onRender" function as props',
+              {withoutStack: true},
+            );
           });
+        }
 
-          it('should render children', () => {
-            const FunctionComponent = ({label}) => <span>{label}</span>;
-            const renderer = ReactTestRenderer.create(
+        it('should support an empty Profiler (with no children)', () => {
+          // As root
+          expect(
+            ReactTestRenderer.create(
+              <React.unstable_Profiler id="label" onRender={jest.fn()} />,
+            ).toJSON(),
+          ).toMatchSnapshot();
+
+          // As non-root
+          expect(
+            ReactTestRenderer.create(
               <div>
-                <span>outside span</span>
-                <React.unstable_Profiler id="label" onRender={jest.fn()}>
-                  <span>inside span</span>
-                  <FunctionComponent label="function component" />
-                </React.unstable_Profiler>
+                <React.unstable_Profiler id="label" onRender={jest.fn()} />
               </div>,
-            );
-            expect(renderer.toJSON()).toMatchSnapshot();
-          });
+            ).toJSON(),
+          ).toMatchSnapshot();
+        });
 
-          it('should support nested Profilers', () => {
-            const FunctionComponent = ({label}) => <div>{label}</div>;
-            class ClassComponent extends React.Component {
-              render() {
-                return <block>{this.props.label}</block>;
-              }
+        it('should render children', () => {
+          const FunctionComponent = ({label}) => <span>{label}</span>;
+          const renderer = ReactTestRenderer.create(
+            <div>
+              <span>outside span</span>
+              <React.unstable_Profiler id="label" onRender={jest.fn()}>
+                <span>inside span</span>
+                <FunctionComponent label="function component" />
+              </React.unstable_Profiler>
+            </div>,
+          );
+          expect(renderer.toJSON()).toMatchSnapshot();
+        });
+
+        it('should support nested Profilers', () => {
+          const FunctionComponent = ({label}) => <div>{label}</div>;
+          class ClassComponent extends React.Component {
+            render() {
+              return <block>{this.props.label}</block>;
             }
-            const renderer = ReactTestRenderer.create(
-              <React.unstable_Profiler id="outer" onRender={jest.fn()}>
-                <FunctionComponent label="outer function component" />
-                <React.unstable_Profiler id="inner" onRender={jest.fn()}>
-                  <ClassComponent label="inner class component" />
-                  <span>inner span</span>
-                </React.unstable_Profiler>
-              </React.unstable_Profiler>,
-            );
-            expect(renderer.toJSON()).toMatchSnapshot();
-          });
+          }
+          const renderer = ReactTestRenderer.create(
+            <React.unstable_Profiler id="outer" onRender={jest.fn()}>
+              <FunctionComponent label="outer function component" />
+              <React.unstable_Profiler id="inner" onRender={jest.fn()}>
+                <ClassComponent label="inner class component" />
+                <span>inner span</span>
+              </React.unstable_Profiler>
+            </React.unstable_Profiler>,
+          );
+          expect(renderer.toJSON()).toMatchSnapshot();
         });
       });
     });
   });
 
-  [true, false].forEach(enableSchedulerTracing => {
-    describe('onRender callback', () => {
-      beforeEach(() => {
-        jest.resetModules();
+  describe('onRender callback', () => {
+    beforeEach(() => {
+      jest.resetModules();
 
-        loadModules({enableSchedulerTracing});
+      loadModules();
+    });
+
+    it('should handle errors thrown', () => {
+      const callback = jest.fn(id => {
+        if (id === 'throw') {
+          throw Error('expected');
+        }
       });
 
-      it('should handle errors thrown', () => {
-        const callback = jest.fn(id => {
-          if (id === 'throw') {
-            throw Error('expected');
-          }
-        });
-
-        let didMount = false;
-        class ClassComponent extends React.Component {
-          componentDidMount() {
-            didMount = true;
-          }
-          render() {
-            return this.props.children;
-          }
+      let didMount = false;
+      class ClassComponent extends React.Component {
+        componentDidMount() {
+          didMount = true;
         }
+        render() {
+          return this.props.children;
+        }
+      }
 
-        // Errors thrown from onRender should not break the commit phase,
-        // Or prevent other lifecycles from being called.
-        expect(() =>
-          ReactTestRenderer.create(
-            <ClassComponent>
-              <React.unstable_Profiler id="do-not-throw" onRender={callback}>
-                <React.unstable_Profiler id="throw" onRender={callback}>
+      // Errors thrown from onRender should not break the commit phase,
+      // Or prevent other lifecycles from being called.
+      expect(() =>
+        ReactTestRenderer.create(
+          <ClassComponent>
+            <React.unstable_Profiler id="do-not-throw" onRender={callback}>
+              <React.unstable_Profiler id="throw" onRender={callback}>
+                <div />
+              </React.unstable_Profiler>
+            </React.unstable_Profiler>
+          </ClassComponent>,
+        ),
+      ).toThrow('expected');
+      expect(didMount).toBe(true);
+      expect(callback).toHaveBeenCalledTimes(2);
+    });
+
+    it('is not invoked until the commit phase', () => {
+      const callback = jest.fn();
+
+      const Yield = ({value}) => {
+        ReactTestRenderer.unstable_yield(value);
+        return null;
+      };
+
+      const renderer = ReactTestRenderer.create(
+        <React.unstable_Profiler id="test" onRender={callback}>
+          <Yield value="first" />
+          <Yield value="last" />
+        </React.unstable_Profiler>,
+        {
+          unstable_isConcurrent: true,
+        },
+      );
+
+      // Times are logged until a render is committed.
+      expect(renderer).toFlushAndYieldThrough(['first']);
+      expect(callback).toHaveBeenCalledTimes(0);
+      expect(renderer).toFlushAndYield(['last']);
+      expect(callback).toHaveBeenCalledTimes(1);
+    });
+
+    it('does not record times for components outside of Profiler tree', () => {
+      ReactTestRenderer.create(
+        <div>
+          <AdvanceTime />
+          <AdvanceTime />
+          <AdvanceTime />
+          <AdvanceTime />
+          <AdvanceTime />
+        </div>,
+      );
+
+      // Should be called two times:
+      // 2. To compute the update expiration time
+      // 3. To record the commit time
+      // No additional calls from ProfilerTimer are expected.
+      expect(mockNow).toHaveBeenCalledTimes(2);
+    });
+
+    it('logs render times for both mount and update', () => {
+      const callback = jest.fn();
+
+      advanceTimeBy(5); // 0 -> 5
+
+      const renderer = ReactTestRenderer.create(
+        <React.unstable_Profiler id="test" onRender={callback}>
+          <AdvanceTime />
+        </React.unstable_Profiler>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(1);
+
+      let [call] = callback.mock.calls;
+
+      expect(call).toHaveLength(7);
+      expect(call[0]).toBe('test');
+      expect(call[1]).toBe('mount');
+      expect(call[2]).toBe(10); // actual time
+      expect(call[3]).toBe(10); // base time
+      expect(call[4]).toBe(5); // start time
+      expect(call[5]).toBe(15); // commit time
+      expect(call[6]).toEqual(new Set()); // interaction events
+
+      callback.mockReset();
+
+      advanceTimeBy(20); // 15 -> 35
+
+      renderer.update(
+        <React.unstable_Profiler id="test" onRender={callback}>
+          <AdvanceTime />
+        </React.unstable_Profiler>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(1);
+
+      [call] = callback.mock.calls;
+
+      expect(call).toHaveLength(7);
+      expect(call[0]).toBe('test');
+      expect(call[1]).toBe('update');
+      expect(call[2]).toBe(10); // actual time
+      expect(call[3]).toBe(10); // base time
+      expect(call[4]).toBe(35); // start time
+      expect(call[5]).toBe(45); // commit time
+      expect(call[6]).toEqual(new Set()); // interaction events
+
+      callback.mockReset();
+
+      advanceTimeBy(20); // 45 -> 65
+
+      renderer.update(
+        <React.unstable_Profiler id="test" onRender={callback}>
+          <AdvanceTime byAmount={4} />
+        </React.unstable_Profiler>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(1);
+
+      [call] = callback.mock.calls;
+
+      expect(call).toHaveLength(7);
+      expect(call[0]).toBe('test');
+      expect(call[1]).toBe('update');
+      expect(call[2]).toBe(4); // actual time
+      expect(call[3]).toBe(4); // base time
+      expect(call[4]).toBe(65); // start time
+      expect(call[5]).toBe(69); // commit time
+      expect(call[6]).toEqual(new Set()); // interaction events
+    });
+
+    it('includes render times of nested Profilers in their parent times', () => {
+      const callback = jest.fn();
+
+      advanceTimeBy(5); // 0 -> 5
+
+      ReactTestRenderer.create(
+        <React.Fragment>
+          <React.unstable_Profiler id="parent" onRender={callback}>
+            <AdvanceTime byAmount={10}>
+              <React.unstable_Profiler id="child" onRender={callback}>
+                <AdvanceTime byAmount={20} />
+              </React.unstable_Profiler>
+            </AdvanceTime>
+          </React.unstable_Profiler>
+        </React.Fragment>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(2);
+
+      // Callbacks bubble (reverse order).
+      const [childCall, parentCall] = callback.mock.calls;
+      expect(childCall[0]).toBe('child');
+      expect(parentCall[0]).toBe('parent');
+
+      // Parent times should include child times
+      expect(childCall[2]).toBe(20); // actual time
+      expect(childCall[3]).toBe(20); // base time
+      expect(childCall[4]).toBe(15); // start time
+      expect(childCall[5]).toBe(35); // commit time
+      expect(parentCall[2]).toBe(30); // actual time
+      expect(parentCall[3]).toBe(30); // base time
+      expect(parentCall[4]).toBe(5); // start time
+      expect(parentCall[5]).toBe(35); // commit time
+    });
+
+    it('traces sibling Profilers separately', () => {
+      const callback = jest.fn();
+
+      advanceTimeBy(5); // 0 -> 5
+
+      ReactTestRenderer.create(
+        <React.Fragment>
+          <React.unstable_Profiler id="first" onRender={callback}>
+            <AdvanceTime byAmount={20} />
+          </React.unstable_Profiler>
+          <React.unstable_Profiler id="second" onRender={callback}>
+            <AdvanceTime byAmount={5} />
+          </React.unstable_Profiler>
+        </React.Fragment>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(2);
+
+      const [firstCall, secondCall] = callback.mock.calls;
+      expect(firstCall[0]).toBe('first');
+      expect(secondCall[0]).toBe('second');
+
+      // Parent times should include child times
+      expect(firstCall[2]).toBe(20); // actual time
+      expect(firstCall[3]).toBe(20); // base time
+      expect(firstCall[4]).toBe(5); // start time
+      expect(firstCall[5]).toBe(30); // commit time
+      expect(secondCall[2]).toBe(5); // actual time
+      expect(secondCall[3]).toBe(5); // base time
+      expect(secondCall[4]).toBe(25); // start time
+      expect(secondCall[5]).toBe(30); // commit time
+    });
+
+    it('does not include time spent outside of profile root', () => {
+      const callback = jest.fn();
+
+      advanceTimeBy(5); // 0 -> 5
+
+      ReactTestRenderer.create(
+        <React.Fragment>
+          <AdvanceTime byAmount={20} />
+          <React.unstable_Profiler id="test" onRender={callback}>
+            <AdvanceTime byAmount={5} />
+          </React.unstable_Profiler>
+          <AdvanceTime byAmount={20} />
+        </React.Fragment>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(1);
+
+      const [call] = callback.mock.calls;
+      expect(call[0]).toBe('test');
+      expect(call[2]).toBe(5); // actual time
+      expect(call[3]).toBe(5); // base time
+      expect(call[4]).toBe(25); // start time
+      expect(call[5]).toBe(50); // commit time
+    });
+
+    it('is not called when blocked by sCU false', () => {
+      const callback = jest.fn();
+
+      let instance;
+      class Updater extends React.Component {
+        state = {};
+        render() {
+          instance = this;
+          return this.props.children;
+        }
+      }
+
+      class Pure extends React.PureComponent {
+        render() {
+          return this.props.children;
+        }
+      }
+
+      const renderer = ReactTestRenderer.create(
+        <React.unstable_Profiler id="outer" onRender={callback}>
+          <Updater>
+            <React.unstable_Profiler id="middle" onRender={callback}>
+              <Pure>
+                <React.unstable_Profiler id="inner" onRender={callback}>
                   <div />
                 </React.unstable_Profiler>
-              </React.unstable_Profiler>
-            </ClassComponent>,
-          ),
-        ).toThrow('expected');
-        expect(didMount).toBe(true);
-        expect(callback).toHaveBeenCalledTimes(2);
+              </Pure>
+            </React.unstable_Profiler>
+          </Updater>
+        </React.unstable_Profiler>,
+      );
+
+      // All profile callbacks are called for initial render
+      expect(callback).toHaveBeenCalledTimes(3);
+
+      callback.mockReset();
+
+      renderer.unstable_flushSync(() => {
+        instance.setState({
+          count: 1,
+        });
       });
 
-      it('is not invoked until the commit phase', () => {
+      // Only call profile updates for paths that have re-rendered
+      // Since "inner" is beneath a pure component, it isn't called
+      expect(callback).toHaveBeenCalledTimes(2);
+      expect(callback.mock.calls[0][0]).toBe('middle');
+      expect(callback.mock.calls[1][0]).toBe('outer');
+    });
+
+    it('decreases actual time but not base time when sCU prevents an update', () => {
+      const callback = jest.fn();
+
+      advanceTimeBy(5); // 0 -> 5
+
+      const renderer = ReactTestRenderer.create(
+        <React.unstable_Profiler id="test" onRender={callback}>
+          <AdvanceTime byAmount={10}>
+            <AdvanceTime byAmount={13} shouldComponentUpdate={false} />
+          </AdvanceTime>
+        </React.unstable_Profiler>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(1);
+
+      advanceTimeBy(30); // 28 -> 58
+
+      renderer.update(
+        <React.unstable_Profiler id="test" onRender={callback}>
+          <AdvanceTime byAmount={4}>
+            <AdvanceTime byAmount={7} shouldComponentUpdate={false} />
+          </AdvanceTime>
+        </React.unstable_Profiler>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(2);
+
+      const [mountCall, updateCall] = callback.mock.calls;
+
+      expect(mountCall[1]).toBe('mount');
+      expect(mountCall[2]).toBe(23); // actual time
+      expect(mountCall[3]).toBe(23); // base time
+      expect(mountCall[4]).toBe(5); // start time
+      expect(mountCall[5]).toBe(28); // commit time
+
+      expect(updateCall[1]).toBe('update');
+      expect(updateCall[2]).toBe(4); // actual time
+      expect(updateCall[3]).toBe(17); // base time
+      expect(updateCall[4]).toBe(58); // start time
+      expect(updateCall[5]).toBe(62); // commit time
+    });
+
+    it('includes time spent in render phase lifecycles', () => {
+      class WithLifecycles extends React.Component {
+        state = {};
+        static getDerivedStateFromProps() {
+          advanceTimeBy(3);
+          return null;
+        }
+        shouldComponentUpdate() {
+          advanceTimeBy(7);
+          return true;
+        }
+        render() {
+          advanceTimeBy(5);
+          return null;
+        }
+      }
+
+      const callback = jest.fn();
+
+      advanceTimeBy(5); // 0 -> 5
+
+      const renderer = ReactTestRenderer.create(
+        <React.unstable_Profiler id="test" onRender={callback}>
+          <WithLifecycles />
+        </React.unstable_Profiler>,
+      );
+
+      advanceTimeBy(15); // 13 -> 28
+
+      renderer.update(
+        <React.unstable_Profiler id="test" onRender={callback}>
+          <WithLifecycles />
+        </React.unstable_Profiler>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(2);
+
+      const [mountCall, updateCall] = callback.mock.calls;
+
+      expect(mountCall[1]).toBe('mount');
+      expect(mountCall[2]).toBe(8); // actual time
+      expect(mountCall[3]).toBe(8); // base time
+      expect(mountCall[4]).toBe(5); // start time
+      expect(mountCall[5]).toBe(13); // commit time
+
+      expect(updateCall[1]).toBe('update');
+      expect(updateCall[2]).toBe(15); // actual time
+      expect(updateCall[3]).toBe(15); // base time
+      expect(updateCall[4]).toBe(28); // start time
+      expect(updateCall[5]).toBe(43); // commit time
+    });
+
+    describe('with regard to interruptions', () => {
+      it('should accumulate actual time after a scheduling interruptions', () => {
         const callback = jest.fn();
 
-        const Yield = ({value}) => {
-          ReactTestRenderer.unstable_yield(value);
+        const Yield = ({renderTime}) => {
+          advanceTimeBy(renderTime);
+          ReactTestRenderer.unstable_yield('Yield:' + renderTime);
           return null;
         };
 
+        advanceTimeBy(5); // 0 -> 5
+
+        // Render partially, but run out of time before completing.
         const renderer = ReactTestRenderer.create(
           <React.unstable_Profiler id="test" onRender={callback}>
-            <Yield value="first" />
-            <Yield value="last" />
+            <Yield renderTime={2} />
+            <Yield renderTime={3} />
           </React.unstable_Profiler>,
-          {
-            unstable_isConcurrent: true,
-          },
+          {unstable_isConcurrent: true},
         );
-
-        // Times are logged until a render is committed.
-        expect(renderer).toFlushAndYieldThrough(['first']);
+        expect(renderer).toFlushAndYieldThrough(['Yield:2']);
         expect(callback).toHaveBeenCalledTimes(0);
-        expect(renderer).toFlushAndYield(['last']);
+
+        // Resume render for remaining children.
+        expect(renderer).toFlushAndYield(['Yield:3']);
+
+        // Verify that logged times include both durations above.
         expect(callback).toHaveBeenCalledTimes(1);
-      });
-
-      it('does not record times for components outside of Profiler tree', () => {
-        ReactTestRenderer.create(
-          <div>
-            <AdvanceTime />
-            <AdvanceTime />
-            <AdvanceTime />
-            <AdvanceTime />
-            <AdvanceTime />
-          </div>,
-        );
-
-        // Should be called two times:
-        // 2. To compute the update expiration time
-        // 3. To record the commit time
-        // No additional calls from ProfilerTimer are expected.
-        expect(mockNow).toHaveBeenCalledTimes(2);
-      });
-
-      it('logs render times for both mount and update', () => {
-        const callback = jest.fn();
-
-        advanceTimeBy(5); // 0 -> 5
-
-        const renderer = ReactTestRenderer.create(
-          <React.unstable_Profiler id="test" onRender={callback}>
-            <AdvanceTime />
-          </React.unstable_Profiler>,
-        );
-
-        expect(callback).toHaveBeenCalledTimes(1);
-
-        let [call] = callback.mock.calls;
-
-        expect(call).toHaveLength(enableSchedulerTracing ? 7 : 6);
-        expect(call[0]).toBe('test');
-        expect(call[1]).toBe('mount');
-        expect(call[2]).toBe(10); // actual time
-        expect(call[3]).toBe(10); // base time
-        expect(call[4]).toBe(5); // start time
-        expect(call[5]).toBe(15); // commit time
-        expect(call[6]).toEqual(enableSchedulerTracing ? new Set() : undefined); // interaction events
-
-        callback.mockReset();
-
-        advanceTimeBy(20); // 15 -> 35
-
-        renderer.update(
-          <React.unstable_Profiler id="test" onRender={callback}>
-            <AdvanceTime />
-          </React.unstable_Profiler>,
-        );
-
-        expect(callback).toHaveBeenCalledTimes(1);
-
-        [call] = callback.mock.calls;
-
-        expect(call).toHaveLength(enableSchedulerTracing ? 7 : 6);
-        expect(call[0]).toBe('test');
-        expect(call[1]).toBe('update');
-        expect(call[2]).toBe(10); // actual time
-        expect(call[3]).toBe(10); // base time
-        expect(call[4]).toBe(35); // start time
-        expect(call[5]).toBe(45); // commit time
-        expect(call[6]).toEqual(enableSchedulerTracing ? new Set() : undefined); // interaction events
-
-        callback.mockReset();
-
-        advanceTimeBy(20); // 45 -> 65
-
-        renderer.update(
-          <React.unstable_Profiler id="test" onRender={callback}>
-            <AdvanceTime byAmount={4} />
-          </React.unstable_Profiler>,
-        );
-
-        expect(callback).toHaveBeenCalledTimes(1);
-
-        [call] = callback.mock.calls;
-
-        expect(call).toHaveLength(enableSchedulerTracing ? 7 : 6);
-        expect(call[0]).toBe('test');
-        expect(call[1]).toBe('update');
-        expect(call[2]).toBe(4); // actual time
-        expect(call[3]).toBe(4); // base time
-        expect(call[4]).toBe(65); // start time
-        expect(call[5]).toBe(69); // commit time
-        expect(call[6]).toEqual(enableSchedulerTracing ? new Set() : undefined); // interaction events
-      });
-
-      it('includes render times of nested Profilers in their parent times', () => {
-        const callback = jest.fn();
-
-        advanceTimeBy(5); // 0 -> 5
-
-        ReactTestRenderer.create(
-          <React.Fragment>
-            <React.unstable_Profiler id="parent" onRender={callback}>
-              <AdvanceTime byAmount={10}>
-                <React.unstable_Profiler id="child" onRender={callback}>
-                  <AdvanceTime byAmount={20} />
-                </React.unstable_Profiler>
-              </AdvanceTime>
-            </React.unstable_Profiler>
-          </React.Fragment>,
-        );
-
-        expect(callback).toHaveBeenCalledTimes(2);
-
-        // Callbacks bubble (reverse order).
-        const [childCall, parentCall] = callback.mock.calls;
-        expect(childCall[0]).toBe('child');
-        expect(parentCall[0]).toBe('parent');
-
-        // Parent times should include child times
-        expect(childCall[2]).toBe(20); // actual time
-        expect(childCall[3]).toBe(20); // base time
-        expect(childCall[4]).toBe(15); // start time
-        expect(childCall[5]).toBe(35); // commit time
-        expect(parentCall[2]).toBe(30); // actual time
-        expect(parentCall[3]).toBe(30); // base time
-        expect(parentCall[4]).toBe(5); // start time
-        expect(parentCall[5]).toBe(35); // commit time
-      });
-
-      it('traces sibling Profilers separately', () => {
-        const callback = jest.fn();
-
-        advanceTimeBy(5); // 0 -> 5
-
-        ReactTestRenderer.create(
-          <React.Fragment>
-            <React.unstable_Profiler id="first" onRender={callback}>
-              <AdvanceTime byAmount={20} />
-            </React.unstable_Profiler>
-            <React.unstable_Profiler id="second" onRender={callback}>
-              <AdvanceTime byAmount={5} />
-            </React.unstable_Profiler>
-          </React.Fragment>,
-        );
-
-        expect(callback).toHaveBeenCalledTimes(2);
-
-        const [firstCall, secondCall] = callback.mock.calls;
-        expect(firstCall[0]).toBe('first');
-        expect(secondCall[0]).toBe('second');
-
-        // Parent times should include child times
-        expect(firstCall[2]).toBe(20); // actual time
-        expect(firstCall[3]).toBe(20); // base time
-        expect(firstCall[4]).toBe(5); // start time
-        expect(firstCall[5]).toBe(30); // commit time
-        expect(secondCall[2]).toBe(5); // actual time
-        expect(secondCall[3]).toBe(5); // base time
-        expect(secondCall[4]).toBe(25); // start time
-        expect(secondCall[5]).toBe(30); // commit time
-      });
-
-      it('does not include time spent outside of profile root', () => {
-        const callback = jest.fn();
-
-        advanceTimeBy(5); // 0 -> 5
-
-        ReactTestRenderer.create(
-          <React.Fragment>
-            <AdvanceTime byAmount={20} />
-            <React.unstable_Profiler id="test" onRender={callback}>
-              <AdvanceTime byAmount={5} />
-            </React.unstable_Profiler>
-            <AdvanceTime byAmount={20} />
-          </React.Fragment>,
-        );
-
-        expect(callback).toHaveBeenCalledTimes(1);
-
         const [call] = callback.mock.calls;
-        expect(call[0]).toBe('test');
         expect(call[2]).toBe(5); // actual time
         expect(call[3]).toBe(5); // base time
-        expect(call[4]).toBe(25); // start time
-        expect(call[5]).toBe(50); // commit time
+        expect(call[4]).toBe(5); // start time
+        expect(call[5]).toBe(10); // commit time
       });
 
-      it('is not called when blocked by sCU false', () => {
+      it('should not include time between frames', () => {
         const callback = jest.fn();
 
-        let instance;
-        class Updater extends React.Component {
-          state = {};
-          render() {
-            instance = this;
-            return this.props.children;
-          }
-        }
+        const Yield = ({renderTime}) => {
+          advanceTimeBy(renderTime);
+          ReactTestRenderer.unstable_yield('Yield:' + renderTime);
+          return null;
+        };
 
-        class Pure extends React.PureComponent {
-          render() {
-            return this.props.children;
-          }
-        }
+        advanceTimeBy(5); // 0 -> 5
 
+        // Render partially, but don't finish.
+        // This partial render should take 5ms of simulated time.
         const renderer = ReactTestRenderer.create(
           <React.unstable_Profiler id="outer" onRender={callback}>
-            <Updater>
-              <React.unstable_Profiler id="middle" onRender={callback}>
-                <Pure>
-                  <React.unstable_Profiler id="inner" onRender={callback}>
-                    <div />
-                  </React.unstable_Profiler>
-                </Pure>
-              </React.unstable_Profiler>
-            </Updater>
+            <Yield renderTime={5} />
+            <Yield renderTime={10} />
+            <React.unstable_Profiler id="inner" onRender={callback}>
+              <Yield renderTime={17} />
+            </React.unstable_Profiler>
           </React.unstable_Profiler>,
+          {unstable_isConcurrent: true},
         );
+        expect(renderer).toFlushAndYieldThrough(['Yield:5']);
+        expect(callback).toHaveBeenCalledTimes(0);
 
-        // All profile callbacks are called for initial render
-        expect(callback).toHaveBeenCalledTimes(3);
+        // Simulate time moving forward while frame is paused.
+        advanceTimeBy(50); // 10 -> 60
+
+        // Flush the remaining work,
+        // Which should take an additional 10ms of simulated time.
+        expect(renderer).toFlushAndYield(['Yield:10', 'Yield:17']);
+        expect(callback).toHaveBeenCalledTimes(2);
+
+        const [innerCall, outerCall] = callback.mock.calls;
+
+        // Verify that the actual time includes all work times,
+        // But not the time that elapsed between frames.
+        expect(innerCall[0]).toBe('inner');
+        expect(innerCall[2]).toBe(17); // actual time
+        expect(innerCall[3]).toBe(17); // base time
+        expect(innerCall[4]).toBe(70); // start time
+        expect(innerCall[5]).toBe(87); // commit time
+        expect(outerCall[0]).toBe('outer');
+        expect(outerCall[2]).toBe(32); // actual time
+        expect(outerCall[3]).toBe(32); // base time
+        expect(outerCall[4]).toBe(5); // start time
+        expect(outerCall[5]).toBe(87); // commit time
+      });
+
+      it('should report the expected times when a high-pri update replaces a mount in-progress', () => {
+        const callback = jest.fn();
+
+        const Yield = ({renderTime}) => {
+          advanceTimeBy(renderTime);
+          ReactTestRenderer.unstable_yield('Yield:' + renderTime);
+          return null;
+        };
+
+        advanceTimeBy(5); // 0 -> 5
+
+        // Render a partially update, but don't finish.
+        // This partial render should take 10ms of simulated time.
+        const renderer = ReactTestRenderer.create(
+          <React.unstable_Profiler id="test" onRender={callback}>
+            <Yield renderTime={10} />
+            <Yield renderTime={20} />
+          </React.unstable_Profiler>,
+          {unstable_isConcurrent: true},
+        );
+        expect(renderer).toFlushAndYieldThrough(['Yield:10']);
+        expect(callback).toHaveBeenCalledTimes(0);
+
+        // Simulate time moving forward while frame is paused.
+        advanceTimeBy(100); // 15 -> 115
+
+        // Interrupt with higher priority work.
+        // The interrupted work simulates an additional 5ms of time.
+        renderer.unstable_flushSync(() => {
+          renderer.update(
+            <React.unstable_Profiler id="test" onRender={callback}>
+              <Yield renderTime={5} />
+            </React.unstable_Profiler>,
+          );
+        });
+        expect(ReactTestRenderer).toHaveYielded(['Yield:5']);
+
+        // The initial work was thrown away in this case,
+        // So the actual and base times should only include the final rendered tree times.
+        expect(callback).toHaveBeenCalledTimes(1);
+        let call = callback.mock.calls[0];
+        expect(call[2]).toBe(5); // actual time
+        expect(call[3]).toBe(5); // base time
+        expect(call[4]).toBe(115); // start time
+        expect(call[5]).toBe(120); // commit time
 
         callback.mockReset();
 
-        renderer.unstable_flushSync(() => {
-          instance.setState({
-            count: 1,
-          });
-        });
-
-        // Only call profile updates for paths that have re-rendered
-        // Since "inner" is beneath a pure component, it isn't called
-        expect(callback).toHaveBeenCalledTimes(2);
-        expect(callback.mock.calls[0][0]).toBe('middle');
-        expect(callback.mock.calls[1][0]).toBe('outer');
+        // Verify no more unexpected callbacks from low priority work
+        expect(renderer).toFlushWithoutYielding();
+        expect(callback).toHaveBeenCalledTimes(0);
       });
 
-      it('decreases actual time but not base time when sCU prevents an update', () => {
+      it('should report the expected times when a high-priority update replaces a low-priority update', () => {
         const callback = jest.fn();
+
+        const Yield = ({renderTime}) => {
+          advanceTimeBy(renderTime);
+          ReactTestRenderer.unstable_yield('Yield:' + renderTime);
+          return null;
+        };
 
         advanceTimeBy(5); // 0 -> 5
 
         const renderer = ReactTestRenderer.create(
           <React.unstable_Profiler id="test" onRender={callback}>
-            <AdvanceTime byAmount={10}>
-              <AdvanceTime byAmount={13} shouldComponentUpdate={false} />
-            </AdvanceTime>
+            <Yield renderTime={6} />
+            <Yield renderTime={15} />
           </React.unstable_Profiler>,
+          {unstable_isConcurrent: true},
         );
 
+        // Render everything initially.
+        // This should take 21 seconds of actual and base time.
+        expect(renderer).toFlushAndYield(['Yield:6', 'Yield:15']);
         expect(callback).toHaveBeenCalledTimes(1);
+        let call = callback.mock.calls[0];
+        expect(call[2]).toBe(21); // actual time
+        expect(call[3]).toBe(21); // base time
+        expect(call[4]).toBe(5); // start time
+        expect(call[5]).toBe(26); // commit time
 
-        advanceTimeBy(30); // 28 -> 58
+        callback.mockReset();
 
+        advanceTimeBy(30); // 26 -> 56
+
+        // Render a partially update, but don't finish.
+        // This partial render should take 3ms of simulated time.
         renderer.update(
           <React.unstable_Profiler id="test" onRender={callback}>
-            <AdvanceTime byAmount={4}>
-              <AdvanceTime byAmount={7} shouldComponentUpdate={false} />
-            </AdvanceTime>
+            <Yield renderTime={3} />
+            <Yield renderTime={5} />
+            <Yield renderTime={9} />
           </React.unstable_Profiler>,
         );
+        expect(renderer).toFlushAndYieldThrough(['Yield:3']);
+        expect(callback).toHaveBeenCalledTimes(0);
 
-        expect(callback).toHaveBeenCalledTimes(2);
+        // Simulate time moving forward while frame is paused.
+        advanceTimeBy(100); // 59 -> 159
 
-        const [mountCall, updateCall] = callback.mock.calls;
+        // Render another 5ms of simulated time.
+        expect(renderer).toFlushAndYieldThrough(['Yield:5']);
+        expect(callback).toHaveBeenCalledTimes(0);
 
-        expect(mountCall[1]).toBe('mount');
-        expect(mountCall[2]).toBe(23); // actual time
-        expect(mountCall[3]).toBe(23); // base time
-        expect(mountCall[4]).toBe(5); // start time
-        expect(mountCall[5]).toBe(28); // commit time
+        // Simulate time moving forward while frame is paused.
+        advanceTimeBy(100); // 164 -> 264
 
-        expect(updateCall[1]).toBe('update');
-        expect(updateCall[2]).toBe(4); // actual time
-        expect(updateCall[3]).toBe(17); // base time
-        expect(updateCall[4]).toBe(58); // start time
-        expect(updateCall[5]).toBe(62); // commit time
-      });
-
-      it('includes time spent in render phase lifecycles', () => {
-        class WithLifecycles extends React.Component {
-          state = {};
-          static getDerivedStateFromProps() {
-            advanceTimeBy(3);
-            return null;
-          }
-          shouldComponentUpdate() {
-            advanceTimeBy(7);
-            return true;
-          }
-          render() {
-            advanceTimeBy(5);
-            return null;
-          }
-        }
-
-        const callback = jest.fn();
-
-        advanceTimeBy(5); // 0 -> 5
-
-        const renderer = ReactTestRenderer.create(
-          <React.unstable_Profiler id="test" onRender={callback}>
-            <WithLifecycles />
-          </React.unstable_Profiler>,
-        );
-
-        advanceTimeBy(15); // 13 -> 28
-
-        renderer.update(
-          <React.unstable_Profiler id="test" onRender={callback}>
-            <WithLifecycles />
-          </React.unstable_Profiler>,
-        );
-
-        expect(callback).toHaveBeenCalledTimes(2);
-
-        const [mountCall, updateCall] = callback.mock.calls;
-
-        expect(mountCall[1]).toBe('mount');
-        expect(mountCall[2]).toBe(8); // actual time
-        expect(mountCall[3]).toBe(8); // base time
-        expect(mountCall[4]).toBe(5); // start time
-        expect(mountCall[5]).toBe(13); // commit time
-
-        expect(updateCall[1]).toBe('update');
-        expect(updateCall[2]).toBe(15); // actual time
-        expect(updateCall[3]).toBe(15); // base time
-        expect(updateCall[4]).toBe(28); // start time
-        expect(updateCall[5]).toBe(43); // commit time
-      });
-
-      describe('with regard to interruptions', () => {
-        it('should accumulate actual time after a scheduling interruptions', () => {
-          const callback = jest.fn();
-
-          const Yield = ({renderTime}) => {
-            advanceTimeBy(renderTime);
-            ReactTestRenderer.unstable_yield('Yield:' + renderTime);
-            return null;
-          };
-
-          advanceTimeBy(5); // 0 -> 5
-
-          // Render partially, but run out of time before completing.
-          const renderer = ReactTestRenderer.create(
-            <React.unstable_Profiler id="test" onRender={callback}>
-              <Yield renderTime={2} />
-              <Yield renderTime={3} />
-            </React.unstable_Profiler>,
-            {unstable_isConcurrent: true},
-          );
-          expect(renderer).toFlushAndYieldThrough(['Yield:2']);
-          expect(callback).toHaveBeenCalledTimes(0);
-
-          // Resume render for remaining children.
-          expect(renderer).toFlushAndYield(['Yield:3']);
-
-          // Verify that logged times include both durations above.
-          expect(callback).toHaveBeenCalledTimes(1);
-          const [call] = callback.mock.calls;
-          expect(call[2]).toBe(5); // actual time
-          expect(call[3]).toBe(5); // base time
-          expect(call[4]).toBe(5); // start time
-          expect(call[5]).toBe(10); // commit time
-        });
-
-        it('should not include time between frames', () => {
-          const callback = jest.fn();
-
-          const Yield = ({renderTime}) => {
-            advanceTimeBy(renderTime);
-            ReactTestRenderer.unstable_yield('Yield:' + renderTime);
-            return null;
-          };
-
-          advanceTimeBy(5); // 0 -> 5
-
-          // Render partially, but don't finish.
-          // This partial render should take 5ms of simulated time.
-          const renderer = ReactTestRenderer.create(
-            <React.unstable_Profiler id="outer" onRender={callback}>
-              <Yield renderTime={5} />
-              <Yield renderTime={10} />
-              <React.unstable_Profiler id="inner" onRender={callback}>
-                <Yield renderTime={17} />
-              </React.unstable_Profiler>
-            </React.unstable_Profiler>,
-            {unstable_isConcurrent: true},
-          );
-          expect(renderer).toFlushAndYieldThrough(['Yield:5']);
-          expect(callback).toHaveBeenCalledTimes(0);
-
-          // Simulate time moving forward while frame is paused.
-          advanceTimeBy(50); // 10 -> 60
-
-          // Flush the remaining work,
-          // Which should take an additional 10ms of simulated time.
-          expect(renderer).toFlushAndYield(['Yield:10', 'Yield:17']);
-          expect(callback).toHaveBeenCalledTimes(2);
-
-          const [innerCall, outerCall] = callback.mock.calls;
-
-          // Verify that the actual time includes all work times,
-          // But not the time that elapsed between frames.
-          expect(innerCall[0]).toBe('inner');
-          expect(innerCall[2]).toBe(17); // actual time
-          expect(innerCall[3]).toBe(17); // base time
-          expect(innerCall[4]).toBe(70); // start time
-          expect(innerCall[5]).toBe(87); // commit time
-          expect(outerCall[0]).toBe('outer');
-          expect(outerCall[2]).toBe(32); // actual time
-          expect(outerCall[3]).toBe(32); // base time
-          expect(outerCall[4]).toBe(5); // start time
-          expect(outerCall[5]).toBe(87); // commit time
-        });
-
-        it('should report the expected times when a high-pri update replaces a mount in-progress', () => {
-          const callback = jest.fn();
-
-          const Yield = ({renderTime}) => {
-            advanceTimeBy(renderTime);
-            ReactTestRenderer.unstable_yield('Yield:' + renderTime);
-            return null;
-          };
-
-          advanceTimeBy(5); // 0 -> 5
-
-          // Render a partially update, but don't finish.
-          // This partial render should take 10ms of simulated time.
-          const renderer = ReactTestRenderer.create(
-            <React.unstable_Profiler id="test" onRender={callback}>
-              <Yield renderTime={10} />
-              <Yield renderTime={20} />
-            </React.unstable_Profiler>,
-            {unstable_isConcurrent: true},
-          );
-          expect(renderer).toFlushAndYieldThrough(['Yield:10']);
-          expect(callback).toHaveBeenCalledTimes(0);
-
-          // Simulate time moving forward while frame is paused.
-          advanceTimeBy(100); // 15 -> 115
-
-          // Interrupt with higher priority work.
-          // The interrupted work simulates an additional 5ms of time.
-          renderer.unstable_flushSync(() => {
-            renderer.update(
-              <React.unstable_Profiler id="test" onRender={callback}>
-                <Yield renderTime={5} />
-              </React.unstable_Profiler>,
-            );
-          });
-          expect(ReactTestRenderer).toHaveYielded(['Yield:5']);
-
-          // The initial work was thrown away in this case,
-          // So the actual and base times should only include the final rendered tree times.
-          expect(callback).toHaveBeenCalledTimes(1);
-          let call = callback.mock.calls[0];
-          expect(call[2]).toBe(5); // actual time
-          expect(call[3]).toBe(5); // base time
-          expect(call[4]).toBe(115); // start time
-          expect(call[5]).toBe(120); // commit time
-
-          callback.mockReset();
-
-          // Verify no more unexpected callbacks from low priority work
-          expect(renderer).toFlushWithoutYielding();
-          expect(callback).toHaveBeenCalledTimes(0);
-        });
-
-        it('should report the expected times when a high-priority update replaces a low-priority update', () => {
-          const callback = jest.fn();
-
-          const Yield = ({renderTime}) => {
-            advanceTimeBy(renderTime);
-            ReactTestRenderer.unstable_yield('Yield:' + renderTime);
-            return null;
-          };
-
-          advanceTimeBy(5); // 0 -> 5
-
-          const renderer = ReactTestRenderer.create(
-            <React.unstable_Profiler id="test" onRender={callback}>
-              <Yield renderTime={6} />
-              <Yield renderTime={15} />
-            </React.unstable_Profiler>,
-            {unstable_isConcurrent: true},
-          );
-
-          // Render everything initially.
-          // This should take 21 seconds of actual and base time.
-          expect(renderer).toFlushAndYield(['Yield:6', 'Yield:15']);
-          expect(callback).toHaveBeenCalledTimes(1);
-          let call = callback.mock.calls[0];
-          expect(call[2]).toBe(21); // actual time
-          expect(call[3]).toBe(21); // base time
-          expect(call[4]).toBe(5); // start time
-          expect(call[5]).toBe(26); // commit time
-
-          callback.mockReset();
-
-          advanceTimeBy(30); // 26 -> 56
-
-          // Render a partially update, but don't finish.
-          // This partial render should take 3ms of simulated time.
+        // Interrupt with higher priority work.
+        // The interrupted work simulates an additional 11ms of time.
+        renderer.unstable_flushSync(() => {
           renderer.update(
             <React.unstable_Profiler id="test" onRender={callback}>
-              <Yield renderTime={3} />
-              <Yield renderTime={5} />
-              <Yield renderTime={9} />
+              <Yield renderTime={11} />
             </React.unstable_Profiler>,
           );
-          expect(renderer).toFlushAndYieldThrough(['Yield:3']);
-          expect(callback).toHaveBeenCalledTimes(0);
-
-          // Simulate time moving forward while frame is paused.
-          advanceTimeBy(100); // 59 -> 159
-
-          // Render another 5ms of simulated time.
-          expect(renderer).toFlushAndYieldThrough(['Yield:5']);
-          expect(callback).toHaveBeenCalledTimes(0);
-
-          // Simulate time moving forward while frame is paused.
-          advanceTimeBy(100); // 164 -> 264
-
-          // Interrupt with higher priority work.
-          // The interrupted work simulates an additional 11ms of time.
-          renderer.unstable_flushSync(() => {
-            renderer.update(
-              <React.unstable_Profiler id="test" onRender={callback}>
-                <Yield renderTime={11} />
-              </React.unstable_Profiler>,
-            );
-          });
-          expect(ReactTestRenderer).toHaveYielded(['Yield:11']);
-
-          // The actual time should include only the most recent render,
-          // Because this lets us avoid a lot of commit phase reset complexity.
-          // The base time includes only the final rendered tree times.
-          expect(callback).toHaveBeenCalledTimes(1);
-          call = callback.mock.calls[0];
-          expect(call[2]).toBe(11); // actual time
-          expect(call[3]).toBe(11); // base time
-          expect(call[4]).toBe(264); // start time
-          expect(call[5]).toBe(275); // commit time
-
-          // Verify no more unexpected callbacks from low priority work
-          expect(renderer).toFlushAndYield([]);
-          expect(callback).toHaveBeenCalledTimes(1);
         });
+        expect(ReactTestRenderer).toHaveYielded(['Yield:11']);
 
-        it('should report the expected times when a high-priority update interrupts a low-priority update', () => {
-          const callback = jest.fn();
+        // The actual time should include only the most recent render,
+        // Because this lets us avoid a lot of commit phase reset complexity.
+        // The base time includes only the final rendered tree times.
+        expect(callback).toHaveBeenCalledTimes(1);
+        call = callback.mock.calls[0];
+        expect(call[2]).toBe(11); // actual time
+        expect(call[3]).toBe(11); // base time
+        expect(call[4]).toBe(264); // start time
+        expect(call[5]).toBe(275); // commit time
 
-          const Yield = ({renderTime}) => {
-            advanceTimeBy(renderTime);
-            ReactTestRenderer.unstable_yield('Yield:' + renderTime);
-            return null;
-          };
-
-          let first;
-          class FirstComponent extends React.Component {
-            state = {renderTime: 1};
-            render() {
-              first = this;
-              advanceTimeBy(this.state.renderTime);
-              ReactTestRenderer.unstable_yield(
-                'FirstComponent:' + this.state.renderTime,
-              );
-              return <Yield renderTime={4} />;
-            }
-          }
-          let second;
-          class SecondComponent extends React.Component {
-            state = {renderTime: 2};
-            render() {
-              second = this;
-              advanceTimeBy(this.state.renderTime);
-              ReactTestRenderer.unstable_yield(
-                'SecondComponent:' + this.state.renderTime,
-              );
-              return <Yield renderTime={7} />;
-            }
-          }
-
-          advanceTimeBy(5); // 0 -> 5
-
-          const renderer = ReactTestRenderer.create(
-            <React.unstable_Profiler id="test" onRender={callback}>
-              <FirstComponent />
-              <SecondComponent />
-            </React.unstable_Profiler>,
-            {unstable_isConcurrent: true},
-          );
-
-          // Render everything initially.
-          // This simulates a total of 14ms of actual render time.
-          // The base render time is also 14ms for the initial render.
-          expect(renderer).toFlushAndYield([
-            'FirstComponent:1',
-            'Yield:4',
-            'SecondComponent:2',
-            'Yield:7',
-          ]);
-          expect(callback).toHaveBeenCalledTimes(1);
-          let call = callback.mock.calls[0];
-          expect(call[2]).toBe(14); // actual time
-          expect(call[3]).toBe(14); // base time
-          expect(call[4]).toBe(5); // start time
-          expect(call[5]).toBe(19); // commit time
-
-          callback.mockClear();
-
-          advanceTimeBy(100); // 19 -> 119
-
-          // Render a partially update, but don't finish.
-          // This partial render will take 10ms of actual render time.
-          first.setState({renderTime: 10});
-          expect(renderer).toFlushAndYieldThrough(['FirstComponent:10']);
-          expect(callback).toHaveBeenCalledTimes(0);
-
-          // Simulate time moving forward while frame is paused.
-          advanceTimeBy(100); // 129 -> 229
-
-          // Interrupt with higher priority work.
-          // This simulates a total of 37ms of actual render time.
-          renderer.unstable_flushSync(() => second.setState({renderTime: 30}));
-          expect(ReactTestRenderer).toHaveYielded([
-            'SecondComponent:30',
-            'Yield:7',
-          ]);
-
-          // The actual time should include only the most recent render (37ms),
-          // Because this greatly simplifies the commit phase logic.
-          // The base time should include the more recent times for the SecondComponent subtree,
-          // As well as the original times for the FirstComponent subtree.
-          expect(callback).toHaveBeenCalledTimes(1);
-          call = callback.mock.calls[0];
-          expect(call[2]).toBe(37); // actual time
-          expect(call[3]).toBe(42); // base time
-          expect(call[4]).toBe(229); // start time
-          expect(call[5]).toBe(266); // commit time
-
-          callback.mockClear();
-
-          // Simulate time moving forward while frame is paused.
-          advanceTimeBy(100); // 266 -> 366
-
-          // Resume the original low priority update, with rebased state.
-          // This simulates a total of 14ms of actual render time,
-          // And does not include the original (interrupted) 10ms.
-          // The tree contains 42ms of base render time at this point,
-          // Reflecting the most recent (longer) render durations.
-          // TODO: This actual time should decrease by 10ms once the scheduler supports resuming.
-          expect(renderer).toFlushAndYield(['FirstComponent:10', 'Yield:4']);
-          expect(callback).toHaveBeenCalledTimes(1);
-          call = callback.mock.calls[0];
-          expect(call[2]).toBe(14); // actual time
-          expect(call[3]).toBe(51); // base time
-          expect(call[4]).toBe(366); // start time
-          expect(call[5]).toBe(380); // commit time
-        });
-
-        [true, false].forEach(
-          replayFailedUnitOfWorkWithInvokeGuardedCallback => {
-            describe(`replayFailedUnitOfWorkWithInvokeGuardedCallback ${
-              replayFailedUnitOfWorkWithInvokeGuardedCallback
-                ? 'enabled'
-                : 'disabled'
-            }`, () => {
-              beforeEach(() => {
-                jest.resetModules();
-
-                loadModules({
-                  replayFailedUnitOfWorkWithInvokeGuardedCallback,
-                });
-              });
-
-              it('should accumulate actual time after an error handled by componentDidCatch()', () => {
-                const callback = jest.fn();
-
-                const ThrowsError = () => {
-                  advanceTimeBy(3);
-                  throw Error('expected error');
-                };
-
-                class ErrorBoundary extends React.Component {
-                  state = {error: null};
-                  componentDidCatch(error) {
-                    this.setState({error});
-                  }
-                  render() {
-                    advanceTimeBy(2);
-                    return this.state.error === null ? (
-                      this.props.children
-                    ) : (
-                      <AdvanceTime byAmount={20} />
-                    );
-                  }
-                }
-
-                advanceTimeBy(5); // 0 -> 5
-
-                ReactTestRenderer.create(
-                  <React.unstable_Profiler id="test" onRender={callback}>
-                    <ErrorBoundary>
-                      <AdvanceTime byAmount={9} />
-                      <ThrowsError />
-                    </ErrorBoundary>
-                  </React.unstable_Profiler>,
-                );
-
-                expect(callback).toHaveBeenCalledTimes(2);
-
-                // Callbacks bubble (reverse order).
-                let [mountCall, updateCall] = callback.mock.calls;
-
-                // The initial mount only includes the ErrorBoundary (which takes 2)
-                // But it spends time rendering all of the failed subtree also.
-                expect(mountCall[1]).toBe('mount');
-                // actual time includes: 2 (ErrorBoundary) + 9 (AdvanceTime) + 3 (ThrowsError)
-                // We don't count the time spent in replaying the failed unit of work (ThrowsError)
-                expect(mountCall[2]).toBe(14);
-                // base time includes: 2 (ErrorBoundary)
-                // Since the tree is empty for the initial commit
-                expect(mountCall[3]).toBe(2);
-                // start time
-                expect(mountCall[4]).toBe(5);
-                // commit time: 5 initially + 14 of work
-                // Add an additional 3 (ThrowsError) if we replayed the failed work
-                expect(mountCall[5]).toBe(
-                  __DEV__ && replayFailedUnitOfWorkWithInvokeGuardedCallback
-                    ? 22
-                    : 19,
-                );
-
-                // The update includes the ErrorBoundary and its fallback child
-                expect(updateCall[1]).toBe('update');
-                // actual time includes: 2 (ErrorBoundary) + 20 (AdvanceTime)
-                expect(updateCall[2]).toBe(22);
-                // base time includes: 2 (ErrorBoundary) + 20 (AdvanceTime)
-                expect(updateCall[3]).toBe(22);
-                // start time
-                expect(updateCall[4]).toBe(
-                  __DEV__ && replayFailedUnitOfWorkWithInvokeGuardedCallback
-                    ? 22
-                    : 19,
-                );
-                // commit time: 19 (startTime) + 2 (ErrorBoundary) + 20 (AdvanceTime)
-                // Add an additional 3 (ThrowsError) if we replayed the failed work
-                expect(updateCall[5]).toBe(
-                  __DEV__ && replayFailedUnitOfWorkWithInvokeGuardedCallback
-                    ? 44
-                    : 41,
-                );
-              });
-
-              it('should accumulate actual time after an error handled by getDerivedStateFromError()', () => {
-                const callback = jest.fn();
-
-                const ThrowsError = () => {
-                  advanceTimeBy(10);
-                  throw Error('expected error');
-                };
-
-                class ErrorBoundary extends React.Component {
-                  state = {error: null};
-                  static getDerivedStateFromError(error) {
-                    return {error};
-                  }
-                  render() {
-                    advanceTimeBy(2);
-                    return this.state.error === null ? (
-                      this.props.children
-                    ) : (
-                      <AdvanceTime byAmount={20} />
-                    );
-                  }
-                }
-
-                advanceTimeBy(5); // 0 -> 5
-
-                ReactTestRenderer.create(
-                  <React.unstable_Profiler id="test" onRender={callback}>
-                    <ErrorBoundary>
-                      <AdvanceTime byAmount={5} />
-                      <ThrowsError />
-                    </ErrorBoundary>
-                  </React.unstable_Profiler>,
-                );
-
-                expect(callback).toHaveBeenCalledTimes(1);
-
-                // Callbacks bubble (reverse order).
-                let [mountCall] = callback.mock.calls;
-
-                // The initial mount includes the ErrorBoundary's error state,
-                // But it also spends actual time rendering UI that fails and isn't included.
-                expect(mountCall[1]).toBe('mount');
-                // actual time includes: 2 (ErrorBoundary) + 5 (AdvanceTime) + 10 (ThrowsError)
-                // Then the re-render: 2 (ErrorBoundary) + 20 (AdvanceTime)
-                // We don't count the time spent in replaying the failed unit of work (ThrowsError)
-                expect(mountCall[2]).toBe(39);
-                // base time includes: 2 (ErrorBoundary) + 20 (AdvanceTime)
-                expect(mountCall[3]).toBe(22);
-                // start time
-                expect(mountCall[4]).toBe(5);
-                // commit time
-                expect(mountCall[5]).toBe(
-                  __DEV__ && replayFailedUnitOfWorkWithInvokeGuardedCallback
-                    ? 54
-                    : 44,
-                );
-              });
-
-              it('should reset the fiber stack correct after a "complete" phase error', () => {
-                jest.resetModules();
-
-                loadModules({
-                  useNoopRenderer: true,
-                  replayFailedUnitOfWorkWithInvokeGuardedCallback,
-                });
-
-                // Simulate a renderer error during the "complete" phase.
-                // This mimics behavior like React Native's View/Text nesting validation.
-                ReactNoop.render(
-                  <React.unstable_Profiler id="profiler" onRender={jest.fn()}>
-                    <errorInCompletePhase>hi</errorInCompletePhase>
-                  </React.unstable_Profiler>,
-                );
-                expect(ReactNoop.flush).toThrow('Error in host config.');
-
-                // A similar case we've seen caused by an invariant in ReactDOM.
-                // It didn't reproduce without a host component inside.
-                ReactNoop.render(
-                  <React.unstable_Profiler id="profiler" onRender={jest.fn()}>
-                    <errorInCompletePhase>
-                      <span>hi</span>
-                    </errorInCompletePhase>
-                  </React.unstable_Profiler>,
-                );
-                expect(ReactNoop.flush).toThrow('Error in host config.');
-
-                // So long as the profiler timer's fiber stack is reset correctly,
-                // Subsequent renders should not error.
-                ReactNoop.render(
-                  <React.unstable_Profiler id="profiler" onRender={jest.fn()}>
-                    <span>hi</span>
-                  </React.unstable_Profiler>,
-                );
-                ReactNoop.flush();
-              });
-            });
-          },
-        );
+        // Verify no more unexpected callbacks from low priority work
+        expect(renderer).toFlushAndYield([]);
+        expect(callback).toHaveBeenCalledTimes(1);
       });
 
-      it('reflects the most recently rendered id value', () => {
+      it('should report the expected times when a high-priority update interrupts a low-priority update', () => {
         const callback = jest.fn();
+
+        const Yield = ({renderTime}) => {
+          advanceTimeBy(renderTime);
+          ReactTestRenderer.unstable_yield('Yield:' + renderTime);
+          return null;
+        };
+
+        let first;
+        class FirstComponent extends React.Component {
+          state = {renderTime: 1};
+          render() {
+            first = this;
+            advanceTimeBy(this.state.renderTime);
+            ReactTestRenderer.unstable_yield(
+              'FirstComponent:' + this.state.renderTime,
+            );
+            return <Yield renderTime={4} />;
+          }
+        }
+        let second;
+        class SecondComponent extends React.Component {
+          state = {renderTime: 2};
+          render() {
+            second = this;
+            advanceTimeBy(this.state.renderTime);
+            ReactTestRenderer.unstable_yield(
+              'SecondComponent:' + this.state.renderTime,
+            );
+            return <Yield renderTime={7} />;
+          }
+        }
 
         advanceTimeBy(5); // 0 -> 5
 
         const renderer = ReactTestRenderer.create(
-          <React.unstable_Profiler id="one" onRender={callback}>
-            <AdvanceTime byAmount={2} />
-          </React.unstable_Profiler>,
-        );
-
-        expect(callback).toHaveBeenCalledTimes(1);
-
-        advanceTimeBy(20); // 7 -> 27
-
-        renderer.update(
-          <React.unstable_Profiler id="two" onRender={callback}>
-            <AdvanceTime byAmount={1} />
-          </React.unstable_Profiler>,
-        );
-
-        expect(callback).toHaveBeenCalledTimes(2);
-
-        const [mountCall, updateCall] = callback.mock.calls;
-
-        expect(mountCall[0]).toBe('one');
-        expect(mountCall[1]).toBe('mount');
-        expect(mountCall[2]).toBe(2); // actual time
-        expect(mountCall[3]).toBe(2); // base time
-        expect(mountCall[4]).toBe(5); // start time
-
-        expect(updateCall[0]).toBe('two');
-        expect(updateCall[1]).toBe('update');
-        expect(updateCall[2]).toBe(1); // actual time
-        expect(updateCall[3]).toBe(1); // base time
-        expect(updateCall[4]).toBe(27); // start time
-      });
-
-      it('should not be called until after mutations', () => {
-        let classComponentMounted = false;
-        const callback = jest.fn(
-          (id, phase, actualDuration, baseDuration, startTime, commitTime) => {
-            // Don't call this hook until after mutations
-            expect(classComponentMounted).toBe(true);
-            // But the commit time should reflect pre-mutation
-            expect(commitTime).toBe(2);
-          },
-        );
-
-        class ClassComponent extends React.Component {
-          componentDidMount() {
-            advanceTimeBy(5);
-            classComponentMounted = true;
-          }
-          render() {
-            advanceTimeBy(2);
-            return null;
-          }
-        }
-
-        ReactTestRenderer.create(
           <React.unstable_Profiler id="test" onRender={callback}>
-            <ClassComponent />
+            <FirstComponent />
+            <SecondComponent />
           </React.unstable_Profiler>,
+          {unstable_isConcurrent: true},
         );
 
+        // Render everything initially.
+        // This simulates a total of 14ms of actual render time.
+        // The base render time is also 14ms for the initial render.
+        expect(renderer).toFlushAndYield([
+          'FirstComponent:1',
+          'Yield:4',
+          'SecondComponent:2',
+          'Yield:7',
+        ]);
         expect(callback).toHaveBeenCalledTimes(1);
+        let call = callback.mock.calls[0];
+        expect(call[2]).toBe(14); // actual time
+        expect(call[3]).toBe(14); // base time
+        expect(call[4]).toBe(5); // start time
+        expect(call[5]).toBe(19); // commit time
+
+        callback.mockClear();
+
+        advanceTimeBy(100); // 19 -> 119
+
+        // Render a partially update, but don't finish.
+        // This partial render will take 10ms of actual render time.
+        first.setState({renderTime: 10});
+        expect(renderer).toFlushAndYieldThrough(['FirstComponent:10']);
+        expect(callback).toHaveBeenCalledTimes(0);
+
+        // Simulate time moving forward while frame is paused.
+        advanceTimeBy(100); // 129 -> 229
+
+        // Interrupt with higher priority work.
+        // This simulates a total of 37ms of actual render time.
+        renderer.unstable_flushSync(() => second.setState({renderTime: 30}));
+        expect(ReactTestRenderer).toHaveYielded([
+          'SecondComponent:30',
+          'Yield:7',
+        ]);
+
+        // The actual time should include only the most recent render (37ms),
+        // Because this greatly simplifies the commit phase logic.
+        // The base time should include the more recent times for the SecondComponent subtree,
+        // As well as the original times for the FirstComponent subtree.
+        expect(callback).toHaveBeenCalledTimes(1);
+        call = callback.mock.calls[0];
+        expect(call[2]).toBe(37); // actual time
+        expect(call[3]).toBe(42); // base time
+        expect(call[4]).toBe(229); // start time
+        expect(call[5]).toBe(266); // commit time
+
+        callback.mockClear();
+
+        // Simulate time moving forward while frame is paused.
+        advanceTimeBy(100); // 266 -> 366
+
+        // Resume the original low priority update, with rebased state.
+        // This simulates a total of 14ms of actual render time,
+        // And does not include the original (interrupted) 10ms.
+        // The tree contains 42ms of base render time at this point,
+        // Reflecting the most recent (longer) render durations.
+        // TODO: This actual time should decrease by 10ms once the scheduler supports resuming.
+        expect(renderer).toFlushAndYield(['FirstComponent:10', 'Yield:4']);
+        expect(callback).toHaveBeenCalledTimes(1);
+        call = callback.mock.calls[0];
+        expect(call[2]).toBe(14); // actual time
+        expect(call[3]).toBe(51); // base time
+        expect(call[4]).toBe(366); // start time
+        expect(call[5]).toBe(380); // commit time
       });
+
+      [true, false].forEach(replayFailedUnitOfWorkWithInvokeGuardedCallback => {
+        describe(`replayFailedUnitOfWorkWithInvokeGuardedCallback ${
+          replayFailedUnitOfWorkWithInvokeGuardedCallback
+            ? 'enabled'
+            : 'disabled'
+        }`, () => {
+          beforeEach(() => {
+            jest.resetModules();
+
+            loadModules({
+              replayFailedUnitOfWorkWithInvokeGuardedCallback,
+            });
+          });
+
+          it('should accumulate actual time after an error handled by componentDidCatch()', () => {
+            const callback = jest.fn();
+
+            const ThrowsError = () => {
+              advanceTimeBy(3);
+              throw Error('expected error');
+            };
+
+            class ErrorBoundary extends React.Component {
+              state = {error: null};
+              componentDidCatch(error) {
+                this.setState({error});
+              }
+              render() {
+                advanceTimeBy(2);
+                return this.state.error === null ? (
+                  this.props.children
+                ) : (
+                  <AdvanceTime byAmount={20} />
+                );
+              }
+            }
+
+            advanceTimeBy(5); // 0 -> 5
+
+            ReactTestRenderer.create(
+              <React.unstable_Profiler id="test" onRender={callback}>
+                <ErrorBoundary>
+                  <AdvanceTime byAmount={9} />
+                  <ThrowsError />
+                </ErrorBoundary>
+              </React.unstable_Profiler>,
+            );
+
+            expect(callback).toHaveBeenCalledTimes(2);
+
+            // Callbacks bubble (reverse order).
+            let [mountCall, updateCall] = callback.mock.calls;
+
+            // The initial mount only includes the ErrorBoundary (which takes 2)
+            // But it spends time rendering all of the failed subtree also.
+            expect(mountCall[1]).toBe('mount');
+            // actual time includes: 2 (ErrorBoundary) + 9 (AdvanceTime) + 3 (ThrowsError)
+            // We don't count the time spent in replaying the failed unit of work (ThrowsError)
+            expect(mountCall[2]).toBe(14);
+            // base time includes: 2 (ErrorBoundary)
+            // Since the tree is empty for the initial commit
+            expect(mountCall[3]).toBe(2);
+            // start time
+            expect(mountCall[4]).toBe(5);
+            // commit time: 5 initially + 14 of work
+            // Add an additional 3 (ThrowsError) if we replayed the failed work
+            expect(mountCall[5]).toBe(
+              __DEV__ && replayFailedUnitOfWorkWithInvokeGuardedCallback
+                ? 22
+                : 19,
+            );
+
+            // The update includes the ErrorBoundary and its fallback child
+            expect(updateCall[1]).toBe('update');
+            // actual time includes: 2 (ErrorBoundary) + 20 (AdvanceTime)
+            expect(updateCall[2]).toBe(22);
+            // base time includes: 2 (ErrorBoundary) + 20 (AdvanceTime)
+            expect(updateCall[3]).toBe(22);
+            // start time
+            expect(updateCall[4]).toBe(
+              __DEV__ && replayFailedUnitOfWorkWithInvokeGuardedCallback
+                ? 22
+                : 19,
+            );
+            // commit time: 19 (startTime) + 2 (ErrorBoundary) + 20 (AdvanceTime)
+            // Add an additional 3 (ThrowsError) if we replayed the failed work
+            expect(updateCall[5]).toBe(
+              __DEV__ && replayFailedUnitOfWorkWithInvokeGuardedCallback
+                ? 44
+                : 41,
+            );
+          });
+
+          it('should accumulate actual time after an error handled by getDerivedStateFromError()', () => {
+            const callback = jest.fn();
+
+            const ThrowsError = () => {
+              advanceTimeBy(10);
+              throw Error('expected error');
+            };
+
+            class ErrorBoundary extends React.Component {
+              state = {error: null};
+              static getDerivedStateFromError(error) {
+                return {error};
+              }
+              render() {
+                advanceTimeBy(2);
+                return this.state.error === null ? (
+                  this.props.children
+                ) : (
+                  <AdvanceTime byAmount={20} />
+                );
+              }
+            }
+
+            advanceTimeBy(5); // 0 -> 5
+
+            ReactTestRenderer.create(
+              <React.unstable_Profiler id="test" onRender={callback}>
+                <ErrorBoundary>
+                  <AdvanceTime byAmount={5} />
+                  <ThrowsError />
+                </ErrorBoundary>
+              </React.unstable_Profiler>,
+            );
+
+            expect(callback).toHaveBeenCalledTimes(1);
+
+            // Callbacks bubble (reverse order).
+            let [mountCall] = callback.mock.calls;
+
+            // The initial mount includes the ErrorBoundary's error state,
+            // But it also spends actual time rendering UI that fails and isn't included.
+            expect(mountCall[1]).toBe('mount');
+            // actual time includes: 2 (ErrorBoundary) + 5 (AdvanceTime) + 10 (ThrowsError)
+            // Then the re-render: 2 (ErrorBoundary) + 20 (AdvanceTime)
+            // We don't count the time spent in replaying the failed unit of work (ThrowsError)
+            expect(mountCall[2]).toBe(39);
+            // base time includes: 2 (ErrorBoundary) + 20 (AdvanceTime)
+            expect(mountCall[3]).toBe(22);
+            // start time
+            expect(mountCall[4]).toBe(5);
+            // commit time
+            expect(mountCall[5]).toBe(
+              __DEV__ && replayFailedUnitOfWorkWithInvokeGuardedCallback
+                ? 54
+                : 44,
+            );
+          });
+
+          it('should reset the fiber stack correct after a "complete" phase error', () => {
+            jest.resetModules();
+
+            loadModules({
+              useNoopRenderer: true,
+              replayFailedUnitOfWorkWithInvokeGuardedCallback,
+            });
+
+            // Simulate a renderer error during the "complete" phase.
+            // This mimics behavior like React Native's View/Text nesting validation.
+            ReactNoop.render(
+              <React.unstable_Profiler id="profiler" onRender={jest.fn()}>
+                <errorInCompletePhase>hi</errorInCompletePhase>
+              </React.unstable_Profiler>,
+            );
+            expect(ReactNoop.flush).toThrow('Error in host config.');
+
+            // A similar case we've seen caused by an invariant in ReactDOM.
+            // It didn't reproduce without a host component inside.
+            ReactNoop.render(
+              <React.unstable_Profiler id="profiler" onRender={jest.fn()}>
+                <errorInCompletePhase>
+                  <span>hi</span>
+                </errorInCompletePhase>
+              </React.unstable_Profiler>,
+            );
+            expect(ReactNoop.flush).toThrow('Error in host config.');
+
+            // So long as the profiler timer's fiber stack is reset correctly,
+            // Subsequent renders should not error.
+            ReactNoop.render(
+              <React.unstable_Profiler id="profiler" onRender={jest.fn()}>
+                <span>hi</span>
+              </React.unstable_Profiler>,
+            );
+            ReactNoop.flush();
+          });
+        });
+      });
+    });
+
+    it('reflects the most recently rendered id value', () => {
+      const callback = jest.fn();
+
+      advanceTimeBy(5); // 0 -> 5
+
+      const renderer = ReactTestRenderer.create(
+        <React.unstable_Profiler id="one" onRender={callback}>
+          <AdvanceTime byAmount={2} />
+        </React.unstable_Profiler>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(1);
+
+      advanceTimeBy(20); // 7 -> 27
+
+      renderer.update(
+        <React.unstable_Profiler id="two" onRender={callback}>
+          <AdvanceTime byAmount={1} />
+        </React.unstable_Profiler>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(2);
+
+      const [mountCall, updateCall] = callback.mock.calls;
+
+      expect(mountCall[0]).toBe('one');
+      expect(mountCall[1]).toBe('mount');
+      expect(mountCall[2]).toBe(2); // actual time
+      expect(mountCall[3]).toBe(2); // base time
+      expect(mountCall[4]).toBe(5); // start time
+
+      expect(updateCall[0]).toBe('two');
+      expect(updateCall[1]).toBe('update');
+      expect(updateCall[2]).toBe(1); // actual time
+      expect(updateCall[3]).toBe(1); // base time
+      expect(updateCall[4]).toBe(27); // start time
+    });
+
+    it('should not be called until after mutations', () => {
+      let classComponentMounted = false;
+      const callback = jest.fn(
+        (id, phase, actualDuration, baseDuration, startTime, commitTime) => {
+          // Don't call this hook until after mutations
+          expect(classComponentMounted).toBe(true);
+          // But the commit time should reflect pre-mutation
+          expect(commitTime).toBe(2);
+        },
+      );
+
+      class ClassComponent extends React.Component {
+        componentDidMount() {
+          advanceTimeBy(5);
+          classComponentMounted = true;
+        }
+        render() {
+          advanceTimeBy(2);
+          return null;
+        }
+      }
+
+      ReactTestRenderer.create(
+        <React.unstable_Profiler id="test" onRender={callback}>
+          <ClassComponent />
+        </React.unstable_Profiler>,
+      );
+
+      expect(callback).toHaveBeenCalledTimes(1);
     });
   });
 
@@ -1277,10 +1267,7 @@ describe('Profiler', () => {
     function loadModulesForTracing(params) {
       jest.resetModules();
 
-      loadModules({
-        enableSchedulerTracing: true,
-        ...params,
-      });
+      loadModules(params);
 
       throwInOnInteractionScheduledWorkCompleted = false;
       throwInOnWorkScheduled = false;
@@ -1555,9 +1542,7 @@ describe('Profiler', () => {
       let call = onRender.mock.calls[0];
       expect(call[0]).toEqual('test-profiler');
       expect(call[5]).toEqual(mockNow());
-      if (ReactFeatureFlags.enableSchedulerTracing) {
-        expect(call[6]).toMatchInteractions([interactionCreation]);
-      }
+      expect(call[6]).toMatchInteractions([interactionCreation]);
 
       expect(onInteractionTraced).toHaveBeenCalledTimes(1);
       expect(onInteractionScheduledWorkCompleted).toHaveBeenCalledTimes(1);
@@ -1627,9 +1612,7 @@ describe('Profiler', () => {
         call = onRender.mock.calls[0];
         expect(call[0]).toEqual('test-profiler');
         expect(call[5]).toEqual(mockNow());
-        if (ReactFeatureFlags.enableSchedulerTracing) {
-          expect(call[6]).toMatchInteractions([interactionOne]);
-        }
+        expect(call[6]).toMatchInteractions([interactionOne]);
 
         didRunCallback = true;
 
@@ -1660,9 +1643,7 @@ describe('Profiler', () => {
       call = onRender.mock.calls[0];
       expect(call[0]).toEqual('test-profiler');
       expect(call[5]).toEqual(mockNow());
-      if (ReactFeatureFlags.enableSchedulerTracing) {
-        expect(call[6]).toMatchInteractions([]);
-      }
+      expect(call[6]).toMatchInteractions([]);
 
       expect(onInteractionTraced).toHaveBeenCalledTimes(2);
       expect(onInteractionScheduledWorkCompleted).toHaveBeenCalledTimes(2);
@@ -1716,9 +1697,7 @@ describe('Profiler', () => {
       call = onRender.mock.calls[0];
       expect(call[0]).toEqual('test-profiler');
       expect(call[5]).toEqual(mockNow());
-      if (ReactFeatureFlags.enableSchedulerTracing) {
-        expect(call[6]).toMatchInteractions([interactionTwo]);
-      }
+      expect(call[6]).toMatchInteractions([interactionTwo]);
 
       expect(onInteractionTraced).toHaveBeenCalledTimes(3);
       expect(onInteractionScheduledWorkCompleted).toHaveBeenCalledTimes(3);
@@ -1849,11 +1828,10 @@ describe('Profiler', () => {
         let call = onRender.mock.calls[0];
         expect(call[0]).toEqual('test');
         expect(call[5]).toEqual(mockNow());
-        expect(call[6]).toMatchInteractions(
-          ReactFeatureFlags.enableSchedulerTracing
-            ? [interactionLowPri, interactionHighPri]
-            : [],
-        );
+        expect(call[6]).toMatchInteractions([
+          interactionLowPri,
+          interactionHighPri,
+        ]);
 
         onRender.mockClear();
 
@@ -1866,9 +1844,7 @@ describe('Profiler', () => {
         call = onRender.mock.calls[0];
         expect(call[0]).toEqual('test');
         expect(call[5]).toEqual(mockNow());
-        expect(call[6]).toMatchInteractions(
-          ReactFeatureFlags.enableSchedulerTracing ? [interactionLowPri] : [],
-        );
+        expect(call[6]).toMatchInteractions([interactionLowPri]);
 
         expect(onInteractionTraced).toHaveBeenCalledTimes(2);
         expect(onInteractionScheduledWorkCompleted).toHaveBeenCalledTimes(1);
@@ -1976,15 +1952,11 @@ describe('Profiler', () => {
       let call = onRender.mock.calls[0];
       expect(call[0]).toEqual('test');
       expect(call[5]).toEqual(firstCommitTime);
-      expect(call[6]).toMatchInteractions(
-        ReactFeatureFlags.enableSchedulerTracing ? [interactionOne] : [],
-      );
+      expect(call[6]).toMatchInteractions([interactionOne]);
       call = onRender.mock.calls[1];
       expect(call[0]).toEqual('test');
       expect(call[5]).toEqual(mockNow());
-      expect(call[6]).toMatchInteractions(
-        ReactFeatureFlags.enableSchedulerTracing ? [interactionOne] : [],
-      );
+      expect(call[6]).toMatchInteractions([interactionOne]);
 
       onRender.mockClear();
 
@@ -2039,15 +2011,11 @@ describe('Profiler', () => {
       call = onRender.mock.calls[0];
       expect(call[0]).toEqual('test');
       expect(call[5]).toEqual(firstCommitTime);
-      expect(call[6]).toMatchInteractions(
-        ReactFeatureFlags.enableSchedulerTracing ? [interactionTwo] : [],
-      );
+      expect(call[6]).toMatchInteractions([interactionTwo]);
       call = onRender.mock.calls[1];
       expect(call[0]).toEqual('test');
       expect(call[5]).toEqual(mockNow());
-      expect(call[6]).toMatchInteractions(
-        ReactFeatureFlags.enableSchedulerTracing ? [interactionTwo] : [],
-      );
+      expect(call[6]).toMatchInteractions([interactionTwo]);
 
       onRender.mockClear();
 
@@ -2104,15 +2072,11 @@ describe('Profiler', () => {
       call = onRender.mock.calls[0];
       expect(call[0]).toEqual('test');
       expect(call[5]).toEqual(firstCommitTime);
-      expect(call[6]).toMatchInteractions(
-        ReactFeatureFlags.enableSchedulerTracing ? [interactionThree] : [],
-      );
+      expect(call[6]).toMatchInteractions([interactionThree]);
       call = onRender.mock.calls[1];
       expect(call[0]).toEqual('test');
       expect(call[5]).toEqual(mockNow());
-      expect(call[6]).toMatchInteractions(
-        ReactFeatureFlags.enableSchedulerTracing ? [interactionThree] : [],
-      );
+      expect(call[6]).toMatchInteractions([interactionThree]);
     });
 
     it('should trace interactions associated with a parent component state update', () => {
@@ -2171,9 +2135,7 @@ describe('Profiler', () => {
       expect(onRender).toHaveBeenCalledTimes(1);
       let call = onRender.mock.calls[0];
       expect(call[0]).toEqual('test-profiler');
-      expect(call[6]).toMatchInteractions(
-        ReactFeatureFlags.enableSchedulerTracing ? [interaction] : [],
-      );
+      expect(call[6]).toMatchInteractions([interaction]);
 
       expect(onInteractionTraced).toHaveBeenCalledTimes(1);
       expect(onInteractionScheduledWorkCompleted).toHaveBeenCalledTimes(1);
@@ -2259,9 +2221,7 @@ describe('Profiler', () => {
 
         let call = onRender.mock.calls[0];
         expect(call[0]).toEqual('test-profiler');
-        expect(call[6]).toMatchInteractions(
-          ReactFeatureFlags.enableSchedulerTracing ? [interaction] : [],
-        );
+        expect(call[6]).toMatchInteractions([interaction]);
 
         expect(onInteractionTraced).toHaveBeenCalledTimes(1);
         expect(onInteractionScheduledWorkCompleted).not.toHaveBeenCalled();
@@ -2282,9 +2242,7 @@ describe('Profiler', () => {
 
         call = onRender.mock.calls[2];
         expect(call[0]).toEqual('test-profiler');
-        expect(call[6]).toMatchInteractions(
-          ReactFeatureFlags.enableSchedulerTracing ? [interaction] : [],
-        );
+        expect(call[6]).toMatchInteractions([interaction]);
 
         expect(onInteractionTraced).toHaveBeenCalledTimes(1);
         expect(onInteractionScheduledWorkCompleted).toHaveBeenCalledTimes(1);
