+++ packages/react/src/__tests__/ReactProfilerDOM-test.internal.js
@@ -46,7 +46,6 @@ function loadModules() {
   ReactFeatureFlags.debugRenderPhaseSideEffects = false;
   ReactFeatureFlags.debugRenderPhaseSideEffectsForStrictMode = false;
   ReactFeatureFlags.enableProfilerTimer = true;
-  ReactFeatureFlags.enableSchedulerTracing = true;
 
   React = require('react');
   SchedulerTracing = require('scheduler/tracing');
