+++ packages/shared/ReactFeatureFlags.js
@@ -29,9 +29,6 @@ export const warnAboutDeprecatedLifecycles = false;
 // Gather advanced timing metrics for Profiler subtrees.
 export const enableProfilerTimer = __PROFILE__;
 
-// Trace which interactions trigger each commit.
-export const enableSchedulerTracing = __PROFILE__;
-
 // Only used in www builds.
 export const enableSuspenseServerRenderer = false; // TODO: __DEV__? Here it might just be false.
