+++ packages/shared/forks/ReactFeatureFlags.native-fb.js
@@ -19,7 +19,6 @@ export const {debugRenderPhaseSideEffects} = require('ReactFeatureFlags');
 export const enableHooks = true;
 export const enableUserTimingAPI = __DEV__;
 export const enableProfilerTimer = __PROFILE__;
-export const enableSchedulerTracing = __PROFILE__;
 export const enableSuspenseServerRenderer = false;
 export const enableStableConcurrentModeAPIs = false;
 export const warnAboutShorthandPropertyCollision = false;
