+++ packages/shared/forks/ReactFeatureFlags.www.js
@@ -33,7 +33,6 @@ export const enableHooks = true;
 export let enableUserTimingAPI = __DEV__;
 
 export const enableProfilerTimer = __PROFILE__;
-export const enableSchedulerTracing = __PROFILE__;
 export const enableSchedulerDebugging = __DEV__; // TODO or just true
 
 export const enableStableConcurrentModeAPIs = false;
